Allow passing extra options with and without "--"
We want to be able to use "--" to pass extra arguments to the
underlying `scp` command.  The current command line parser assumes
that the command is in one of two forms:

    govuk-connect [command] [args...] [extra...]

Or

    govuk-connect [command] [args...] -- [extra...]

Where the `args` are determined by the command.

For `scp` we want to be able to copy around arbitrary numbers of
files (not just a single file at a time).  This somewhat works with
the current parser:

1. If a "--" is given, any arguments after it are passed to the
   command.

2. If a "--" is not given, any extra arguments are passed to the
   command.

However, in case (1), any extra arguments *before* the "--" are thrown
away.  This means we can't simultaneously pass a list of files and a
list of extra arguments to `scp`.

This commit treats the two variadic argument lists differently.
Commands which previously threw an error when given extra arguments
still do; `ssh` just concatenates the lists; and `scp` will treat them
differently.