Add `scp-push` and `scp-pull` commands
Each take a list of paths (at least two in length), where the last
paths is the destination and the others are the sources

`scp-push` copies the sources from the current machine to the
destination path on the target machine; `scp-pull` does the opposite,
downloading paths from the target machine to the path on the local
machine.

This builds on the previous commit by using the extra
arguments *before* a "--" as the file list and the extra
arguments *after* a "--" as extra `scp` options.